% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity_functions}
\alias{similarity_functions}
\alias{sms_tas_similarity}
\alias{sms_phenotypic_similarity}
\alias{sms_chemical_similarity}
\title{Small Molecule Suite (SMS) compound similarity functions}
\usage{
sms_tas_similarity(query_ids, target_ids, min_n = 6)

sms_phenotypic_similarity(query_id, min_n = 6)

sms_chemical_similarity(query_id)
}
\arguments{
\item{query_ids}{Compounds (\code{lspci_ids}) for which to calculate
similarity. If no target_ids are provided, the similarity of
all query compounds with each other are calculated.}

\item{target_ids}{\code{lspci_ids} of compounds. If provided, all pairwise
similarities between query and target compounds are calculated.}
}
\description{
TAS asdf

Phenotypic asdf

Computes the Tanimoto similarity between Morgan fingerprints.
}
