% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sms_chemical_similarity}
\alias{sms_chemical_similarity}
\title{Calculate chemical similarity}
\usage{
sms_chemical_similarity(query_ids, target_ids = NULL)
}
\arguments{
\item{querys}{Compounds (names or \code{lspci_ids}) for which to calculate
similarity. If no targets are provided, the similarity of
all query compounds with each other are calculated.}

\item{targets}{If provided (names or \code{lspci_ids}), all pairwise
similarities between query and target compounds are calculated.}
}
\value{
A 3 column data frame with query_id, target_id, and similarity
}
\description{
Computes the Tanimoto similarity between Morgan fingerprints.
}
\details{
Before similarity functions can be used, compound data needs
to be downloaded to the working directory from
\href{https://www.synapse.org/#!Synapse:syn25955270}{Synapse} or
automatically using the \code{\link[=sms_download]{sms_download()}} function.

Compound names can be provided directly to the function,
mapped to \code{lspci_ids} using the \code{\link[=sms_compound_ids]{sms_compound_ids()}}
function, or using the \code{lsp_compound_names} table from
\url{https://labsyspharm.shinyapps.io/smallmoleculesuite}.
}
\seealso{
\code{\link[=sms_download]{sms_download()}} for downloading SMS data

Other similarity functions: 
\code{\link{sms_phenotypic_similarity}()},
\code{\link{sms_tas_similarity}()}
}
\concept{similarity functions}
