% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sms_download}
\alias{sms_download}
\title{Download Small Molecule Suite data}
\usage{
sms_download()
}
\description{
Download SMS compound data from
\href{https://www.synapse.org/#!Synapse:syn25955270}{Synapse} to the working
directory. Synapse login credentials must be saved using \code{\link[synapser:synLogin]{synapser::synLogin()}}.
}
\details{
\code{synLogin(email = "xxx", password = "xxx", rememberMe = TRUE)}.
}
