% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sms_phenotypic_similarity}
\alias{sms_phenotypic_similarity}
\title{Calculate phenotypic similarity}
\usage{
sms_phenotypic_similarity(query_id, min_n = 6)
}
\arguments{
\item{query_ids}{Compounds (\code{lspci_ids}) for which to calculate
similarity. If no target_ids are provided, the similarity of
all query compounds with each other are calculated.}

\item{target_ids}{\code{lspci_ids} of compounds. If provided, all pairwise
similarities between query and target compounds are calculated.}
}
\value{
A 3 column data frame with query_id, target_id, and similarity
}
\description{
Computes similarity between compounds based on phenotypic assays performed
on them that are published in ChEMBL. The similarity metric is the Pearson
correlation between the normalized results of all assays shared by both
compounds.
}
\details{
Before similarity functions can be used, compound data needs
to be downloaded to the working directory from
\href{https://www.synapse.org/#!Synapse:syn25955270}{Synapse} or
automatically using the \code{\link[=sms_download]{sms_download()}} function.

Compound names can be mapped to \code{lspci_ids} using the \code{lsp_compound_names}
table from \url{https://labsyspharm.shinyapps.io/smallmoleculesuite}.
}
\seealso{
\code{\link[=sms_download]{sms_download()}} for downloading SMS data

Other similarity functions: 
\code{\link{sms_chemical_similarity}()},
\code{\link{sms_tas_similarity}()}
}
\concept{similarity functions}
